Moved components used on pages to their own files, importing components dynamically where possible.
